Harden hook preview redaction
